+++ changelog.md
@@ -29,6 +29,7 @@ that is decrypted and stored in device.
 - Save in ecdsa key/cert directly in `DER`, no need for `openssl` in `PATH`.
 - Allow mdns name in ssl certificate.
 - Improved file io operations (`put`, `get`, `fget`, `dsync`, `rsync`, `backup`...), more flexibility indicating files / dirs / pattern matching etc. Improved overall performance, specially for WebSocketDevices.
+- fix `ping`, `probe` for zerotier devices (this needs updated config, see how-to section in docs.)
 ## [0.3.7] 2021-12-16
 ## Added
 - `rssi` command in shell repls to get RSSI value (Wifi or Ble)